requirements.txt
bidict==0.23.1
blinker==1.8.2
cffi==1.17.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
setuptools==72.1.0
simple-websocket==1.0.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.1
